__init__.py to gnpy.py
